app/src/main/java/com/ahr/storyapp/ui/detail/DetailViewModel.kt
package com.ahr.storyapp.ui.detail

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.ahr.storyapp.domain.Response
import com.ahr.storyapp.domain.Story
import com.ahr.storyapp.domain.repository.StoryRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import javax.inject.Inject

@HiltViewModel
class DetailViewModel @Inject constructor(private val storyRepository: StoryRepository) : ViewModel() {

    fun getDetailStory(token: String, id: String): LiveData<Response<Story>> {
        return storyRepository.getDetailStory(token, id)
    }

}